requirements.txt
crewai
crewai-tools
langchain-groq
chromadb
streamlit
pymupdf
firecrawl-py
python-dotenv
requests